requirements.txt
python-dotenv
google-genai
fastapi
uvicorn
pydantic
psycopg2-binary
flask
agno
